+++ docs/CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.2 (unreleased)
 ----------------
 
+- For the link_redirect_view, don't modify the urls for ``mailto:``, ``tel:``,
+  ``callto:`` and ``file:`` link types but render them as-is.
+  [thet]
+
 - When installing the default profile, restrict uninstalling of old types to
   old FTI based ones.
   [thet]